feat(web-support) Made demo-components compatible with web. added rnweb environment to test the react-native-mutliple-modals package in an expo/web environment. Added metro config and symlinks to locally develop packages in the rn80 folder.
