+++ .gitignore
@@ -68,7 +68,7 @@ e2e/results/
 e2e/screenshots/
 e2e/expected-screenshots/
 
-.yarn
+.yarn/cache
 .vscode
 .bundle
