+++ demo-components/src/components/scenario-card/ScenarioCard.tsx
@@ -21,8 +21,6 @@ export const ScenarioCard: FC<ScenarioCardProps> = ({
   const {colors} = useTheme();
   const isLargeScreen = useWindowDimensions().width > 600;
 
-  console.debug(props.testID);
-
   return (
     <Pressable
       {...props}
