+++ demo-components/src/modals/slide/AnimatedSlideModal.tsx
@@ -20,6 +20,7 @@ export const AnimatedSlideModal: FC<AnimatedSlideModalProps> = props => {
 const styles = StyleSheet.create({
   contentContainer: {
     justifyContent: 'flex-end',
+    width: '100%',
   },
   modal: {
     width: '100%',
