+++ rn80/package.json
@@ -14,11 +14,11 @@
     "@react-native/new-app-screen": "0.80.1",
     "@react-navigation/bottom-tabs": "^7.4.2",
     "@react-navigation/native": "^7.1.14",
-    "demo-components": "file:../demo-components",
+    "demo-components": "link:../demo-components",
     "react": "19.1.0",
     "react-native": "0.80.1",
     "react-native-gesture-handler": "^2.27.2",
-    "react-native-multiple-modals": "file:../../react-native-multiple-modals",
+    "react-native-multiple-modals": "link:../../react-native-multiple-modals",
     "react-native-reanimated": "^3.19.0",
     "react-native-safe-area-context": "^5.5.2",
     "react-native-screens": "^4.13.1"
@@ -45,5 +45,6 @@
   },
   "engines": {
     "node": ">=18"
-  }
+  },
+  "packageManager": "yarn@4.9.2"
 }
