+++ rnweb/app/index.tsx
@@ -1,0 +1,5 @@
+import { DemoApp } from 'demo-components';
+
+export default function Index() {
+  return (<DemoApp />);
+}
